refactor interpolation code into own module; also refactor helper parts from volume saga module
