+++ frontend/javascripts/types/distance-transform.d.ts
@@ -1,0 +1,1 @@
+declare module "distance-transform";
